+++ ldap_server.js
@@ -170,7 +170,7 @@ LDAP._createClient = function () {
   client.on('error', function (error) {
     LDAP.error('ldapjs client reported an error: ', error);
     client.destroy();
-    throw new Meteor.Error('ldap-error', 'ldapjs client reported an error', error);
+    // throw new Meteor.Error('ldap-error', 'ldapjs client reported an error', error);
   });
 
   return client;